rename authentication back to challenge
