+++ src/routes/challenge.rs
